+++ client/src/components/App.jsx
@@ -1,10 +1,6 @@
-import React, { Component } from 'react';
-import Home from './HomePage';
+import React from 'react';
+import Home from './Home';
 
-class App extends Component {
-  render() {
-    return <Home />
-  }
-}
+const App = () => <Home />;
 
 export default App;
