+++ package.json
@@ -10,7 +10,7 @@
     "build": "npm-run-all --parallel build:server build:client",
     "build:server": "babel server/src -d server/build",
     "build:client": "webpack --config webpack.prod.js",
-    "test": "npm-run-all test:server",
+    "test": "npm-run-all --parallel test:server test:client",
     "pretest:server": "npm-run-all lint:server migrate:test seed:test",
     "test:client": "cross-env NODE_ENV=test nyc mocha client/tests",
     "test:server": "cross-env NODE_ENV=test nyc mocha server/tests",
